third_party/WebKit/LayoutTests/fast/workers/worker-multi-port-expected.txt

This test checks the various use cases around sending multiple ports through Worker.postMessage

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS worker.postMessage() threw exception TypeError: Failed to execute 'postMessage' on 'Worker': 1 argument required, but only 0 present..
PASS worker.postMessage("null port", [channel3.port1, null, channel3.port2]) threw exception TypeError: Failed to execute 'postMessage' on 'Worker': Value at index 1 is an untransferable 'null' value..
PASS worker.postMessage("notAPort", [channel3.port1, {}, channel3.port2]) threw exception TypeError: Failed to execute 'postMessage' on 'Worker': Value at index 1 does not have a transferable type..
PASS worker.postMessage("notASequence", [{length: 3}]) threw exception TypeError: Failed to execute 'postMessage' on 'Worker': Value at index 0 does not have a transferable type..
PASS successfullyParsed is true

TEST COMPLETE
PASS postMessage() threw exception: TypeError: Failed to execute 'postMessage' on 'DedicatedWorkerGlobalScope': 1 argument required, but only 0 present.
PASS event.ports is non-null and zero length when no port sent
PASS event.ports is non-null and zero length when empty array sent
PASS event.ports contains two ports when two ports sent
PASS event.ports contains two ports when two ports re-sent after error

TEST COMPLETE
